+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Version 1.3.1
 - fix submarine exploding bug if L5S3 level was loaded using 'level' command
 - fix killed glass restoration from a save file
 - add optional hit-sounds in dedicated server
+- add 'map_rest', 'map_next', 'map_prev' commands
 
 Version 1.3.0 (released 2019-05-07)
 -----------------------------------